2nd

+++ hellooboe/app/src/main/cpp/HTS_gstream.c
@@ -108,7 +108,7 @@ HTS_Boolean HTS_GStreamSet_create(HTS_GStreamSet * gss, HTS_PStreamSet * pss, in
          gss->gstream[i].par[j] = (double *) HTS_calloc(gss->gstream[i].static_length, sizeof(double));
    }
    extern short *HTS_wavebuffer;
-   HTS_wavebuffer = gss->gspeech = (short *) HTS_calloc((const size_t) gss->total_nsample, sizeof(short));
+   HTS_wavebuffer = gss->gspeech = (short *) HTS_calloc( gss->total_nsample, sizeof(short));
    extern int32_t buffersize;
    buffersize = gss->total_nsample;
    /* copy generated parameter */
